// frontend/src/wordpress-package/types.ts
export interface ACFBase<A = unknown> {
	acf: A
}

export interface YoastBase {
	yoast_head_json?: YoastHead
}

// ToDo: Omit<WP_REST_API_Post, 'menu_order'>
export type WPPost<A = unknown> = WP_REST_API_Post & ACFBase<A> & YoastBase

export type WPMedia<A = unknown> = WP_REST_API_Attachment &
	ACFBase<A> &
	YoastBase

// ToDo: Fix Omit<> type hinting
export type WPPage<A = unknown> /* Omit< */ = WPPost<A> /* ,
	'categories' | 'sticky' | 'tags'
> */ & { menu_order: number; parent: number }

export type WPTaxonomy<A = unknown> = WP_REST_API_Taxonomy &
	ACFBase<A> &
	YoastBase

export type WPCategory<A = unknown> = WP_REST_API_Category &
	ACFBase<A> &
	YoastBase
// ...
export interface YoastHead {
	title: string
	robots: YoastRobots
	og_locale: string
	og_type: string
	og_title: string
	og_description: string
	og_url: string
	og_site_name: string
	article_modified_time?: Date
	article_published_time: Date
	og_image?: {
		width: number
		height: number
		url: string
		type: string
	}[]
	twitter_card: string
	twitter_misc: YoastTwitterMisc
	schema: YoastSchema
}